fix: killed satan

+++ src/main/kotlin/net/illuc/kontraption/controls/KontraptionSeatedControllingPlayer.kt
@@ -3,10 +3,10 @@ package net.illuc.kontraption.controls
 import net.minecraft.core.Direction
 
 class KontraptionSeatedControllingPlayer(val seatInDirection: Direction) {
-    var forwardImpulse: Float = 0.0f
-    var leftImpulse: Float = 0.0f
-    var upImpulse: Float = 0.0f
-    var pitch: Float = 0.0f
-    var yaw: Float = 0.0f
-    var roll: Float = 0.0f
+    var forwardImpulse: Double = 0.0
+    var leftImpulse: Double = 0.0
+    var upImpulse: Double = 0.0
+    var pitch: Double = 0.0
+    var yaw: Double = 0.0
+    var roll: Double = 0.0
 }
